@@ -13,23 +13,23 @@
         "php": "^8.0",
         "ext-filter": "*",
         "ext-json": "*",
-        "psr/log": "^1.0"
+        "psr/log": "^1.1.4|^2.0|^3.0"
     },
     "require-dev": {
-        "chubbyphp/chubbyphp-container": "^1.0|^2.0",
+        "chubbyphp/chubbyphp-container": "^2.1",
         "chubbyphp/chubbyphp-dev-helper": "dev-master",
         "chubbyphp/chubbyphp-laminas-config-factory": "^1.0",
         "chubbyphp/chubbyphp-mock": "^1.6.1",
-        "doctrine/persistence": "^1.3|^2.0",
-        "infection/infection": "^0.25.3",
-        "php-coveralls/php-coveralls": "^2.5.1",
+        "doctrine/persistence": "^2.3",
+        "infection/infection": "^0.26.5",
+        "php-coveralls/php-coveralls": "^2.5.2",
         "phploc/phploc": "^7.0.2",
         "phpstan/extension-installer": "^1.1",
-        "phpstan/phpstan": "^1.1.2",
-        "phpunit/phpunit": "^9.5.10",
-        "pimple/pimple": "^3.3",
-        "psr/container": "^1.0|^2.0",
-        "symfony/validator": "^5.3.11|^6.0.0"
+        "phpstan/phpstan": "^1.4.8",
+        "phpunit/phpunit": "^9.5.17",
+        "pimple/pimple": "^3.5",
+        "psr/container": "^2.0.2",
+        "symfony/validator": "^5.4.5|^6.0"
     },
     "autoload": {
         "psr-4": { "Chubbyphp\\Validation\\": "src/" }
@@ -38,7 +38,11 @@
         "psr-4": { "Chubbyphp\\Tests\\Validation\\": "tests/" }
     },
     "config": {
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "infection/extension-installer": true,
+            "phpstan/extension-installer": true
+        }
     },
     "extra": {
         "branch-alias": {
@@ -57,7 +61,7 @@
             "@test:cs"
         ],
         "test:cs": "mkdir -p build && PHP_CS_FIXER_IGNORE_ENV=1 vendor/bin/php-cs-fixer fix --dry-run --stop-on-violation --cache-file=build/phpcs.cache",
-        "test:infection": "vendor/bin/infection --threads=$(nproc) --min-msi=90 --verbose --coverage=build/phpunit",
+        "test:infection": "vendor/bin/infection --threads=$(nproc) --min-msi=91 --verbose --coverage=build/phpunit",
         "test:integration": "vendor/bin/phpunit --testsuite=Integration --cache-result-file=build/phpunit/result.cache",
         "test:lint": "mkdir -p build && find src tests -name '*.php' -print0 | xargs -0 -n1 -P$(nproc) php -l | tee build/phplint.log",
         "test:loc": "mkdir -p build && vendor/bin/phploc src | tee build/phploc.log",
