@@ -30,19 +30,19 @@ final class CoordinateArrayConstraint implements ConstraintInterface
             return [new Error(
                 $path,
                 'constraint.coordinatearray.invalidtype',
-                ['type' => \is_object($value) ? \get_class($value) : \gettype($value)]
+                ['type' => get_debug_type($value)]
             )];
         }
 
         if (!isset($value['lat']) || !isset($value['lon'])) {
-            return [new Error($path, 'constraint.coordinatearray.invalidformat', ['value' => json_encode($value)])];
+            return [new Error($path, 'constraint.coordinatearray.invalidformat', ['value' => json_encode($value, JSON_THROW_ON_ERROR)])];
         }
 
         $lat = $value['lat'];
         $lon = $value['lon'];
 
         if ($lat < -90 || $lat > 90 || $lon < -180 || $lon > 180) {
-            return [new Error($path, 'constraint.coordinatearray.invalidvalue', ['value' => json_encode($value)])];
+            return [new Error($path, 'constraint.coordinatearray.invalidvalue', ['value' => json_encode($value, JSON_THROW_ON_ERROR)])];
         }
 
         return [];
