@@ -11,14 +11,8 @@ use Chubbyphp\Validation\ValidatorInterface;
 
 final class CountConstraint implements ConstraintInterface
 {
-    private ?int $min = null;
-
-    private ?int $max = null;
-
-    public function __construct(?int $min = null, ?int $max = null)
+    public function __construct(private ?int $min = null, private ?int $max = null)
     {
-        $this->min = $min;
-        $this->max = $max;
     }
 
     /**
@@ -40,7 +34,7 @@ final class CountConstraint implements ConstraintInterface
             return [new Error(
                 $path,
                 'constraint.count.invalidtype',
-                ['type' => \is_object($value) ? \get_class($value) : \gettype($value)]
+                ['type' => get_debug_type($value)]
             )];
         }
 
