@@ -30,7 +30,7 @@ final class NumericConstraint implements ConstraintInterface
             return [new Error(
                 $path,
                 'constraint.numeric.invalidtype',
-                ['type' => \is_object($value) ? \get_class($value) : \gettype($value)]
+                ['type' => get_debug_type($value)]
             )];
         }
 
