@@ -7,16 +7,10 @@ namespace Chubbyphp\Validation;
 final class ValidatorContext implements ValidatorContextInterface
 {
     /**
-     * @var array<int, string>
-     */
-    private array $groups = [];
-
-    /**
      * @param array<int, string> $groups
      */
-    public function __construct(array $groups = [])
+    public function __construct(private array $groups = [])
     {
-        $this->groups = $groups;
     }
 
     /**
